[dev/trello.com/c/AvS5QpYb] Добавил проверку на текущий кошелек. Небольшой рефакторинг. Отображение ошибки текущей выбранной ноды
